Remove Turbolinks
Turbolinks causes some strange behavior that it's easier not to work
around for now. When submitting a form, if there are errors, the form
will re-render with the errors. This also changes the route to the route
that was POSTed to. Without Turbolinks, the browser will ask for
confirmation to re-send the form when reloading the page, which is the
behavior we want. With Turbolinks, the browser tries to GET that URL,
which in our case does not exist.

We can think about re-adding Turbolinks when this PR[1] is merged.

[1] https://github.com/turbolinks/turbolinks/pull/495