Allow users to create an account and login
Add Clearance[1] to provide login, sign up, session management, etc.

Override some of the Clearance defaults to show errors on the forms and
change the redirects.

Change the default Clearance routes to be more resourceful.

Add default action mailer config so that links in emails work,
particularly for the password reset email.

[1]: https://github.com/thoughtbot/clearance/