Require users to confirm their emaill address
This prevents spammy and fraudulent signups.

When a user signs up, we will send them an email in the background. When
a user logs in, they can request that we resend their confirmation
email.

There is a new Rake task to backfill a confirmation token for existing
users. I plan to manually set confirmed_at for the users we have now,
all of whom I know personally.

An Email Confirmation represents the confirmation token that allows the
user to confirm their account. A Confirmation Email represents the email
that we send them.

In order to more easily display messages to the user, we make use of
Rails flashes, which we show underneath the header.